requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# AI & Communication
openai==1.54.3
twilio==9.3.8
httpx==0.27.2

# Database
SQLAlchemy==2.0.36
alembic==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.10
greenlet==3.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
PyJWT==2.9.0

# Email & Validation
email-validator==2.2.0
resend==2.17.0

# Rate limiting & resilience
slowapi==0.1.9
tenacity==9.0.0

# Observability (Phase 2-4 divine fixes)
prometheus-client==0.21.0
